requirements.txt
# FastAPI and Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching and Queue
redis==5.0.1
kafka-python==2.0.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==42.0.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Orchestration (Temporal)
temporalio==1.5.1

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-prometheus==1.12.0rc1

# Logging
loguru==0.7.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
